requirements.txt
annotated-types==0.7.0
anthropic==0.36.2
anyio==4.6.2.post1
blinker==1.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
distro==1.9.0
dnspython==2.6.1
filelock==3.16.1
Flask==3.0.2
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.3
jiter==0.6.1
MarkupSafe==2.1.5
openai==1.52.0
packaging==24.1
PyJWT==2.8.0
pymongo==4.6.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
tokenizers==0.20.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.1